chore: Tray helpers global refactor
Move every helpers / functions used by the systray applet into a
separate and dedicated `tray_helpers.rs` module.
